generation/DeveloperTools/apicagent/api.py
import requests
from typing import Optional, Dict

def get_user_agent_info(ua: str, toolbench_rapidapi_key: str = '088440d910mshef857391f2fc461p17ae9ejsnaebc918926ff') -> Dict:
    """
    Retrieve parsed information about a user agent string using the GET method.

    Parameters:
    ua (str): The user agent string URL encoded.

    Returns:
    dict: A dictionary containing the parsed information of the user agent.
    """
    url = "https://api.apicagent.com"
    params = {'ua': ua}
    
    response = requests.get(url, params=params)
    try:
        return response.json()
    except Exception as e:
        return {"error": str(e), "response": response.text}